requirements.txt
Django==1.10.6
psycopg2==2.7.4
Pillow==4.0.0
easy_thumbnails==2.3

django-sekizai==0.10.0
django-filer==1.2.7
djangocms-admin-style==0.2.8
git+https://github.com/miracle2k/django-assets
django-braces==1.8.1
django-polymorphic==1.1
django-debug-toolbar
django-extensions
django-redis==4.7
django-simple-history
django-modeltranslation
raven==6.0.0
ipython

celery==4.0

transliterate==1.7.3
cssmin==0.2.0
yuicompressor